backend/requirements.txt
aniso8601==8.1.1
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.8
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
Werkzeug==1.0.1
